templates/mxnet-cu90-1.3.1/requirements.txt
algorithmia>=1.0.0,<2.0
six
numpy
mxnet-cu90==1.3.1